file structure and getStyles typing
